+++ nbxsync/models/__init__.py
@@ -29,3 +29,4 @@ from .zabbixmaintenancetagassignment import *
 
 from .zabbixproblem import *
 from .zabbixevent import *
+from .zabbixtemplaterule import *